Added option to have user text in serif (template text Sans)
With guidelines off, it was sometimes difficult to tell when the
form stopped and the user entry began.  With the document already
a bit cluttered, this was easy to skim over and miss.

Create a \usertext command that is either \textsf or \textrm depending
on a new boolean {serif}.  Allow it to be set true or false
on the card, but by default it will be "serif if no guidelines,
sans if guidelines".