Add two "alternating checkbox" test cards to check against misassociations.
Each should create a checkerboard of set and unset boxes, and each should hit
exactly the ones the other does not.  If two in a row are set, or if the
same box is set in both boxes, there's an issue.

Note: strict alternation is overridden in places where there are sets of two, to
create a checkerboard pattern rather than lines (which could be cut-and-paste
errors).  Those are:
- 1M 4/5
- 1NT Nat/Trf
- direct cuebid
- 1x-X-2NT min/limit raise