Start on anchoring labels and boxes to nodes.
Instead of formatting everything with absolute coordinates, use x.east/
x.west and xspace to tie nodes and labels together.

This will allow, for instance, Gerardo's double-width setup to not have
everything totally out of whack.

Not sure if I'm doing this the right way; I've been tying the text/box
to the label, might be the other way around is better (esp. for box matrices).

VERY temporary commit to transfer for snowbirding.