Add notes to MF-EK-KS card.
Had trouble with publishing markdown, so rewrote notes in LaTeX.  Had
difficulty doing both in the same card holder box, and keeping the references
matched, but realized that I could just make it a second page in the
filled-out template.

Once this is set up and works correctly, replace the "real" cards, and
maybe add a section to the template that shows "how to" put in notes for
others.