Add: usertextsize for larger text
It can be hard to read text on the new card because of the density.
This is especially true with the serif font (which is smaller at the same
point size as the sans-serif).
Added a flag \usertextsize to the template that will move to the next
text size up if "serif" set for user text.
Added a section to the template that allows the user to override this
value (either bigger size or smaller, independent of user text font).
Added upgrade script to add that user-selection section to previously
created cards.