Add: upgrade tools functions and classes.
Upgraders are going down a standard path:
- check to ensure it's a card template
- check to ensure changes haven't been already done
- add lines (by line number or matching text)
- delete lines (by line number or matching text)
- replace lines (ditto)
- replace text (in one place or globally).

Wrote functions to do each of these things, and an abstract class
to "do the thing".  Wrote a StandardUpgrader class that takes a
bunch of change instructions and a template/directory and does
the normal thing.

Now upgrades should simply shoving a list of instructions into the
StandardUpgrader and having it do the right thing.